修改背景

+++ Assets/Scripts/UIManager.cs
@@ -26,7 +26,7 @@ public class UIManager : MonoBehaviour
     public static void UpdateTimeUI(float time)
     {
         int minutes = (int)(time / 60);
-        float seconds = time % 60;
+        int seconds = (int)(time % 60);
         instance.timeText.text = minutes.ToString("00") + ":" + seconds.ToString("00");
     }
 }
